CParser: Fixed issue with ReadString and UTF-8
git-svn-id: svn://ultimatepp.org/upp/trunk@7241 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Core/parser.cpp
@@ -327,7 +327,7 @@ String CParser::ReadOneString(int delim, bool chkend) throw(Error) {
 			}
 		}
 		else {
-			if(*term < ' ') {
+			if((byte)*term < ' ') {
 				if(chkend) {
 					ThrowError("Unterminated string");
 					return result;
